VsoEndpointClient
