+++ src/SAMLUser.ts
@@ -1,14 +1,12 @@
-
 export interface SAMLUser {
-    uuid: string;
-    dateOfBirth?: string;
-    email: string;
-    firstName?: string;
-    gender?: string;
-    lastName?: string;
-    middleName?: string;
-    ssn?: string;
-    edipi?: string;
-    icn?: string;
-  }
-  
+  uuid: string;
+  dateOfBirth?: string;
+  email: string;
+  firstName?: string;
+  gender?: string;
+  lastName?: string;
+  middleName?: string;
+  ssn?: string;
+  edipi?: string;
+  icn?: string;
+}
